docker_process_handler/requirements.txt
boto3
pandas
pillow
